Fixing up the DB more.
I also switched it so that the tokens would be String based (UUID +
other data). This will make it less obvjois whats going on behind the
scenes, and if we ever need to change the API hasn't changed at all.

Also added the scema mapper so that i'm able to use different schema's
for dev and prod

+++ web-service/api-project/src/main/kotlin/tech/crom/service/api/user/UserTokenEndpoint.kt
@@ -27,7 +27,7 @@ open class UserTokenEndpoint @Autowired constructor(
     @LoggedInUserRequired
     @RequestMapping(value = "/{tokenId}", method = arrayOf(RequestMethod.DELETE))
     open fun deleteToken(requestDetails: RequestDetails, @PathVariable("tokenId") tokenId: String): ResponseEntity<*> {
-        tokenManagementApi.invalidateToken(tokenId.toLong(), TokenType.USER)
+        tokenManagementApi.invalidateToken(tokenId, TokenType.USER)
         return ResponseEntity<Any>(HttpStatus.OK)
     }
